src/inference/utils/preprocess.py
from torchvision import transforms
from PIL import Image
import io
import torch


def preprocess_image(image_bytes: bytes):
    transform = transforms.Compose(
        [
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
        ]
    )
    image = Image.open(io.BytesIO(image_bytes)).convert("RGB")
    return transform(image).unsqueeze(0)  # shape: (1, 3, 224, 224)
