src/inference/requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pillow
xgboost==2.0.3
boto3
joblib
